Add chart for telemetry
