+++ src/components/telemetry/TelemetryPanel.css
@@ -1,8 +1,9 @@
 .telemetry-panel {
-  width: 100%;
-  height: 100%
-}
-
-.telemetry-panel__motor {
   display: flex;
+  justify-content: center;
+  position: relative;
+  border-radius: 10px;
+  width: 100%;
+  height: 100%;
+  box-sizing: border-box;
 }
