+++ src/components/telemetry/TelemetryPanel.js
@@ -1,27 +1,10 @@
-import { useSelector } from "react-redux";
-import { selectAllMotorNames, selectMotorCurrentPosition, selectMotorCurrentPower, selectMotorCurrentVelocity } from "../../store/motorsSlice";
-import camelCaseToTitle from "../../util/camelCaseToTitle";
+import MotorTelemetry from "./MotorTelemetry";
 import "./TelemetryPanel.css";
 
 function TelemetryPanel() {
-  const motorNames = useSelector(selectAllMotorNames);
-
   return (
     <div className="telemetry-panel">
-      {motorNames.map(motorName => <MotorTelemetry motorName={motorName} key={motorName} />)}
-    </div>
-  );
-}
-
-function MotorTelemetry({ motorName }) {
-  const power = useSelector(selectMotorCurrentPower(motorName));
-  const position = useSelector(selectMotorCurrentPosition(motorName));
-  const velocity = useSelector(selectMotorCurrentVelocity(motorName));
-  const motorTitle = camelCaseToTitle(motorName);
-
-  return (
-    <div className="telemetry-panel__motor">
-      {motorTitle} power: {Math.round(power)} position: {Math.round(position)} deg velocity: {Math.round(velocity)} deg/s
+      <MotorTelemetry />
     </div>
   );
 }
